Move nx_to_directed to generators
