+++ examples/synthetic_graph_evolution/generators.py
@@ -3,10 +3,31 @@ from typing import Dict, Callable, Collection, Sequence
 import networkx as nx
 import numpy as np
 
-from golem.core.adapter.nx_adapter import nx_to_directed
-
 NumNodes = int
 DiGraphGenerator = Callable[[NumNodes], nx.DiGraph]
+
+
+def nx_to_directed(graph: nx.Graph) -> nx.DiGraph:
+    """Randomly chooses a direction for each edge."""
+    dedges = set()
+    digraph = nx.DiGraph()
+
+    for node, data in graph.nodes(data=True):
+        digraph.add_node(node, **data)
+
+    for u, v, data in graph.edges.data():
+        edge = (u, v)
+        inv_edge = (v, u)
+        if edge in dedges or inv_edge in dedges:
+            continue
+
+        if np.random.default_rng().random() > 0.5:
+            digraph.add_edge(*edge, **data)
+            dedges.add(edge)
+        else:
+            digraph.add_edge(*inv_edge, **data)
+            dedges.add(inv_edge)
+    return digraph
 
 
 graph_generators: Dict[str, DiGraphGenerator] = {
